feat: add start:production task
In Sass.js, a comparison statement is added that allows the Sass to be converted to compressed CSS if the project is in a production environment. Otherwise the Sass will be compiled to expanded CSS.

The start:production script in the package.json file sets the node environment to production, then runs the start task.

main.scss and main.css are also updated as I tested to verify that the Sass compiled to CSS correctly.