feat: added script task
Added a "start:server" script task that runs the watch script AND the server script at the same time. I chose to use the npm-run-all node module versus the parallelshell node module because npm-run-all offers the flexibility to run scripts in sequence AND/OR in at the same time.

Ultimately the start script will need to be updated with npm-run-all and all the scripts we want to run at once. Currently the npm run html script referenced in the start script is throwing an error that prevents the watch and server scripts to run. So I created a temporary "start:server" script.