feat: commiting WIP
I've been working to determine how to pull the place card image into the header of the place card using an inline style which accesses the an image referenced in data, without success. This commit saves work in progress so that I can create a new branch to experiment with a npm module that resizes an image. My goal is to resize all images placed in the image folder to a standard size that can then be accessed directly from a data placed in the html.